+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- More detailed logs in send user error handling.
+- Added the `!naughty` command for owners.
+
 ## [v8.4.1](https://github.com/lexicalunit/spellbot/releases/tag/v8.4.1) - 2022-12-02
 
 ### Changed